Crate kimchi

Modules

alphas

This module implements an abstraction to keep track of the powers of alphas. As a recap, alpha is a challenge sent by the verifier in PLONK, and is used to aggregate multiple constraints into a single polynomial. It is important that different constraints use different powers of alpha, as otherwise they can interact and potentially cancel one another. (The proof is in the use of the Schwartz-Zippel lemma.) As such, we want two properties from this:

bench

circuits

error

This module implements the ProofError type.

index

This module implements Plonk Protocol Index primitive.

plonk_sponge

prover

This module implements prover’s zk-proof primitive.

verifier

This module implements zk-proof batch verifier functionality.