kimchi

Module bench

Structs

BenchmarkCtx

Functions

bench_arguments_dump_into_file

This function can be called before any call to a kimchi verfier, in which case it will serialise kimchi inputs so that they can be reused later for re-testing this particular prover. Used for serialising real mina circuits from ocaml and bindings side.

bench_arguments_from_file

Given a filename with encoded (witness, runtime table, prev rec challenges, constrain system), returns arguments necessary to run a prover.

Type Aliases

BaseSpongePallas

BaseSpongeVesta

ScalarSpongePallas

ScalarSpongeVesta