kimchi::bench

Function bench_arguments_dump_into_file

pub fn bench_arguments_dump_into_file<G: KimchiCurve>(
    cs: &ConstraintSystem<G::ScalarField>,
    witness: &[Vec<G::ScalarField>; 15],
    runtime_tables: &[RuntimeTable<G::ScalarField>],
    prev: &[RecursionChallenge<G>],
)

This function can be called before any call to a kimchi verfier, in which case it will serialise kimchi inputs so that they can be reused later for re-testing this particular prover. Used for serialising real mina circuits from ocaml and bindings side.