kimchi::circuits

Module argument

An argument is simply a number of constraints, which we want to enforce on all points of the domain. Both the permutation and the plookup arguments fit this type. Gates can be seen as filtered arguments, which apply only in some points (rows) of the domain. For more info, read book/src/kimchi/arguments.md

Structs

ArgumentData

Argument environment data for constraints of field elements

ArgumentEnv

The argument environment is used to specify how the argument’s constraints are represented when they are built. If the environment is created without ArgumentData and with F = Expr<F>, then the constraints are built as Expr expressions (e.g. for use with the prover/verifier). On the other hand, if the environment is created with ArgumentData and F = Field or F = PrimeField, then the constraints are built as expressions of real field elements and can be evaluated directly on the witness without using the prover.

ArgumentWitness

Witness data for an argument

Enums

ArgumentType

A constraint type represents a polynomial that will be part of the final equation f (the circuit equation)

Traits

Argument

The interface for a minimal argument implementation.

DynArgument