kimchi::circuits::constraints

Function zk_rows_strict_lower_bound

pub fn zk_rows_strict_lower_bound(num_chunks: usize) -> usize

This function computes a strict lower bound in the number of rows required for zero knowledge in circuits with num_chunks chunks. This means that at least one needs 1 more row than the result of this function to achieve zero knowledge. Example: for 1 chunk, this function returns 2, but at least 3 rows are needed Note: the number of zero knowledge rows is usually computed across the codebase as the formula (16 * num_chunks + 5) / 7, which is precisely the formula in this function plus one.