kimchi::circuits::constraints

Struct ConstraintSystem

pub struct ConstraintSystem<F: PrimeField> {
    pub public: usize,
    pub prev_challenges: usize,
    pub domain: EvaluationDomains<F>,
    pub gates: Arc<Vec<CircuitGate<F>>>,
    pub zk_rows: u64,
    pub feature_flags: FeatureFlags,
    pub sid: Vec<F>,
    pub shift: [F; 7],
    pub endo: F,
    pub lookup_constraint_system: Arc<LazyCache<Result<Option<LookupConstraintSystem<F>>, LookupError>>>,
    pub disable_gates_checks: bool,
    /* private fields */
}

Fields
public: usize

number of public inputs

prev_challenges: usize

number of previous evaluation challenges, for recursive proving

domain: EvaluationDomains<F>

evaluation domains

gates: Arc<Vec<CircuitGate<F>>>

circuit gates

zk_rows: u64feature_flags: FeatureFlags

flags for optional features

sid: Vec<F>

SID polynomial

shift: [F; 7]

wire coordinate shifts

endo: F

coefficient for the group endomorphism

lookup_constraint_system: Arc<LazyCache<Result<Option<LookupConstraintSystem<F>>, LookupError>>>

lookup constraint system

disable_gates_checks: bool

Disable gates checks (for testing; only enables with development builds)

Implementations

impl<F: PrimeField> ConstraintSystem<F>

pub fn create(gates: Vec<CircuitGate<F>>) -> Builder<F>

Initializes the ConstraintSystem<F> on input gates and fr_sponge_params. Returns a Builder<F> It also defaults to the following values of the builder:

public: 0

prev_challenges: 0

lookup_tables: vec![],

runtime_tables: None,

precomputations: None,

disable_gates_checks: false,

lazy_mode: false,

How to use it:

Create your instance of your builder for the constraint system using crate(gates, sponge params)

Iterativelly invoke any desired number of steps: public(), lookup(), runtime(), precomputations(), lazy_mode()

Finally call the build() method and unwrap the Result to obtain your ConstraintSystem

pub fn precomputations(&self) -> Arc<DomainConstantEvaluations<F>>

pub fn for_testing(gates: Vec<CircuitGate<F>>) -> Self

test helpers

pub fn fp_for_testing(gates: Vec<CircuitGate<F>>) -> Self

impl<F: PrimeField> ConstraintSystem<F>

pub fn evaluate(&self, w: &[DP<F>; 15], z: &DP<F>) -> WitnessOverDomains<F>

evaluate witness polynomials over domains

impl<F: PrimeField> ConstraintSystem<F>

pub fn perm_scalars( e: &ProofEvaluations<PointEvaluations<F>>, beta: F, gamma: F, alphas: impl Iterator<Item = F>, zkp_zeta: F, ) -> F