Module kimchi::circuits::expr

Modules

constraints

A number of useful constraints

prologue

You can import this module like use kimchi::circuits::expr::prologue::* to obtain a number of handy aliases and helpers

Macros

auto_clone

Auto clone macro - Helps make constraints more readable by eliminating requirement to .clone() all the time

auto_clone_array

Structs

Cache

A cache

CacheId

A key for a cached value

Challenges

Constants

The collection of constants required to evaluate an Expr.

Environment

The collection of polynomials (all in evaluation form) and constants required to evaluate an expression as a polynomial.

Linearization

A “linearization”, which is linear combination with E coefficients of columns.

LookupEnvironment

The polynomials specific to the lookup argument.

RowOffset

Variable

A type representing a variable which can appear in a constraint. It specifies a column and a relative position (Curr or Next)

Enums

ChallengeTerm

ConstantExprInner

ConstantTerm

Domain

Expr

An multi-variate polynomial over the base ring C with variables

ExprError

FeatureFlag

The feature flags that can be used to enable or disable parts of constraints.

Op2

A binary operation

Operations

PolishToken

For efficiency of evaluation, we compile expressions to reverse Polish notation expressions, which are vectors of the below tokens.

Traits

ColumnEnvironment

ColumnEvaluations

GenericColumn

Literal

ToPolish

Functions

coeff

constant

Convenience function to create a constant as Expr.

index

Handy function to quickly create an expression for a gate.

l0_1

Computes prod_{j != n} (1 - omega^j) Assure we don’t multiply by (1 - omega^n) = (1 - omega^0) = (1 - 1) = 0

pows

Compute the powers of x, x^0, ..., x^{n - 1}

witness

Helper function to quickly create an expression for a witness.

witness_curr

Same as witness but for the current row.

witness_next

Same as witness but for the next row.

Type Aliases

ConstantExpr

E

An alias for the intended usage of the expression type in constructing constraints.