Module kimchi::circuits::expr::constraints

A number of useful constraints

Traits

ArithmeticOps

This trait defines a common arithmetic operations interface that can be used by constraints. It allows us to reuse constraint code for witness computation.

Functions

boolean

Creates a constraint to enforce that b is either 0 or 1.

crumb

Crumb constraint for 2-bit value x