Enum kimchi::circuits::expr::Expr

pub enum Expr<C> {
    Constant(C),
    Cell(Variable),
    Double(Box<Expr<C>>),
    Square(Box<Expr<C>>),
    BinOp(Op2, Box<Expr<C>>, Box<Expr<C>>),
    VanishesOnLast4Rows,
    UnnormalizedLagrangeBasis(i32),
    Pow(Box<Expr<C>>, u64),
    Cache(CacheId, Box<Expr<C>>),
}

An multi-variate polynomial over the base ring C with variables

Cell(v) for v : Variable

VanishesOnLast4Rows

UnnormalizedLagrangeBasis(i) for i : i32

This represents a PLONK “custom constraint”, which enforces that the corresponding combination of the polynomials corresponding to the above variables should vanish on the PLONK domain.

Variants

Constant(C)

Cell(Variable)

Double(Box<Expr<C>>)

Square(Box<Expr<C>>)

BinOp(Op2, Box<Expr<C>>, Box<Expr<C>>)

VanishesOnLast4Rows

UnnormalizedLagrangeBasis(i32)

UnnormalizedLagrangeBasis(i) is (x^n - 1) / (x - omega^i)

Pow(Box<Expr<C>>, u64)

Cache(CacheId, Box<Expr<C>>)

Implementations

impl<C> Expr<C>

pub fn cell(col: Column, row: CurrOrNext) -> Expr<C>

Convenience function for constructing cell variables.

pub fn double(self) -> Self

pub fn square(self) -> Self

pub fn constant(c: C) -> Expr<C>

Convenience function for constructing constant expressions.

impl<F: Field> Expr<ConstantExpr<F>>

pub fn literal(x: F) -> Self

Convenience function for constructing expressions from literal field elements.

pub fn combine_constraints(
alphas: impl Iterator<Item = u32>,
cs: Vec<Self>
) -> Self

Combines multiple constraints [c0, ..., cn] into a single constraint alpha^alpha0 * c0 + alpha^{alpha0 + 1} * c1 + ... + alpha^{alpha0 + n} * cn.

impl<F: FftField> Expr<ConstantExpr<F>>

pub fn to_polish(&self) -> Vec<PolishToken<F>>

Compile an expression to an RPN expression.

pub fn beta() -> Self

The expression beta.

pub fn evaluate(
&self,
d: D<F>,
pt: F,
evals: &[ProofEvaluations<F>],
env: &Environment<'_, F>
) -> Result<F, ExprError>

Evaluate an expression as a field element against an environment.

pub fn evaluate_(
&self,
d: D<F>,
pt: F,
evals: &[ProofEvaluations<F>],
c: &Constants<F>
) -> Result<F, ExprError>

Evaluate an expression as a field element against the constants.

pub fn evaluate_constants(&self, env: &Environment<'_, F>) -> Expr<F>

Evaluate the constant expressions in this expression down into field elements.

pub fn evaluations<'a>(&self, env: &Environment<'a, F>) -> Evaluations<F, D<F>>

Compute the polynomial corresponding to this expression, in evaluation form.

impl<F: FftField> Expr<F>

pub fn evaluate(
&self,
d: D<F>,
pt: F,
evals: &[ProofEvaluations<F>]
) -> Result<F, ExprError>

Evaluate an expression into a field element.

pub fn evaluations<'a>(&self, env: &Environment<'a, F>) -> Evaluations<F, D<F>>

Compute the polynomial corresponding to this expression, in evaluation form.

impl<F: One> Expr<F>

pub fn pow(self, p: u64) -> Self

Exponentiate an expression

impl<F: Neg<Output = F> + Clone + One + Zero + PartialEq> Expr<F>

pub fn linearize(
&self,
evaluated: HashSet<Column>
) -> Result<Linearization<Expr<F>>, ExprError>

There is an optimization in PLONK called “linearization” in which a certain polynomial is expressed as a linear combination of other polynomials in order to reduce the number of evaluations needed in the IOP (by relying on the homomorphic property of the polynomial commitments used.)

The function performs this “linearization”, which we now describe in some detail.

In mathematical language, an expression e: Expr<F> is an element of the polynomial ring F[V], where V is a set of variables.

Given a subset V_0 of V (and letting V_1 = V \setminus V_0), there is a map factor_{V_0}: F[V] -> (F[V_1])[V_0]. That is, polynomials with F coefficients in the variables V = V_0 \cup V_1 are the same thing as polynomials with F[V_1] coefficients in variables V_0.

There is also a function lin_or_err : (F[V_1])[V_0] -> Result<Vec<(V_0, F[V_2])>, &str>

which checks if the given input is in fact a degree 1 polynomial in the variables V_0 (i.e., a linear combination of V_0 elements with F[V_1] coefficients) returning this linear combination if so.

Given an expression e and set of columns C_0, letting V_0 = { Variable { col: c, row: r } | c in C_0, r in { Curr, Next } }, this function computes lin_or_err(factor_{V_0}(e)), although it does not compute it in that way. Instead, it computes it by reducing the expression into a sum of monomials with F coefficients, and then factors the monomials.

impl<F> Expr<ConstantExpr<F>>where
F: PrimeField,

pub fn ocaml_str(&self) -> String

Converts the expression in OCaml code

pub fn latex_str(&self) -> Vec<String>

Converts the expression in LaTeX

pub fn text_str(&self) -> String

Converts the expression to a text string