Struct kimchi::circuits::expr::Environment

pub struct Environment<'a, F: FftField> {
    pub witness: &'a [Evaluations<F, D<F>>; 15],
    pub coefficient: &'a [Evaluations<F, D<F>>; 15],
    pub vanishes_on_last_4_rows: &'a Evaluations<F, D<F>>,
    pub z: &'a Evaluations<F, D<F>>,
    pub index: HashMap<GateType, &'a Evaluations<F, D<F>>>,
    pub l0_1: F,
    pub constants: Constants<F>,
    pub domain: EvaluationDomains<F>,
    pub lookup: Option<LookupEnvironment<'a, F>>,
}

The collection of polynomials (all in evaluation form) and constants required to evaluate an expression as a polynomial.

All are evaluations.

Fields
witness: &'a [Evaluations<F, D<F>>; 15]

The witness column polynomials

coefficient: &'a [Evaluations<F, D<F>>; 15]

The coefficient column polynomials

vanishes_on_last_4_rows: &'a Evaluations<F, D<F>>

The polynomial which vanishes on the last 4 elements of the domain.

z: &'a Evaluations<F, D<F>>

The permutation aggregation polynomial.

index: HashMap<GateType, &'a Evaluations<F, D<F>>>

The index selector polynomials.

l0_1: F

The value prod_{j != 1} (1 - omega^j), used for efficiently computing the evaluations of the unnormalized Lagrange basis polynomials.

constants: Constants<F>

Constant values required

domain: EvaluationDomains<F>

The domains used in the PLONK argument.

lookup: Option<LookupEnvironment<'a, F>>

Lookup specific polynomials