kimchi::circuits

Module gate

This module implements Plonk constraint gate primitive.

Modules

caml

Structs

Circuit

A circuit is specified as a public input size and a list of CircuitGate.

CircuitGate

A single gate in a circuit.

Enums

CircuitGateError

Gate error

CurrOrNext

A row accessible from a given row, corresponds to the fact that we open all polynomials at zeta and omega * zeta.

GateType

The different types of gates the system supports. Note that all the gates are mutually exclusive: they cannot be used at the same time on single row. If we were ever to support this feature, we would have to make sure not to re-use powers of alpha across constraints.

Traits

Connect

Trait to connect a pair of cells in a circuit

Type Aliases

CircuitGateResult

Gate result