Struct kimchi::circuits::gate::CircuitGate

pub struct CircuitGate<F: PrimeField> {
    pub typ: GateType,
    pub wires: GateWires,
    pub coeffs: Vec<F>,
}

A single gate in a circuit.

Fields
typ: GateType

type of the gate

wires: GateWires

gate wiring (for each cell, what cell it is wired to)

coeffs: Vec<F>

public selector polynomials that can used as handy coefficients in gates

Implementations

impl<F> CircuitGate<F>where
F: PrimeField,

pub fn new(typ: GateType, wires: GateWires, coeffs: Vec<F>) -> Self

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn zero(wires: GateWires) -> Self

this function creates “empty” circuit gate

pub fn verify<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15],
index: &ProverIndex<G>,
public: &[F]
) -> Result<(), String>

This function verifies the consistency of the wire assignments (witness) against the constraints

Errors

Will give error if verify process returns error.

pub fn verify_witness<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>,
_public: &[F]
) -> CircuitGateResult<()>

Verify the witness against the constraints

impl<F: PrimeField> CircuitGate<F>

pub fn verify_complete_add(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

Check the correctness of witness values for a complete-add gate.

Errors

Will give error if the gate value validations are not met.

Panics

Will panic if multiplicative inverse operation between gate values fails.

impl<F: PrimeField> CircuitGate<F>

pub fn verify_endomul_scalar<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15],
_cs: &ConstraintSystem<F>
) -> Result<(), String>

Verify the EndoMulscalar gate.

Errors

Will give error if self.typ is not GateType::EndoMulScalar, or there are errors in gate values.

impl<F: PrimeField> CircuitGate<F>

Implementation of group endomorphism optimised variable base scalar multiplication custom Plonk constraints.

pub fn create_endomul(wires: GateWires) -> Self

pub fn verify_endomul<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

Verify the EndoMul gate.

Errors

Will give error if self.typ is not GateType::EndoMul, or constraint evaluation fails.

pub fn endomul(&self) -> F

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_foreign_field_add(
start_row: usize,
num: usize
) -> (usize, Vec<Self>)

Create foreign field addition gate Inputs starting row number of addition gates Outputs tuple (next_row, circuit_gates) where next_row - next row after this gate circuit_gates - vector of circuit gates comprising this gate

Note that te final structure of the circuit is as follows: circuit_gates = [ [0..3] -> 1 RangeCheck for left_input { [8i+4..8i+7] -> 1 RangeCheck for right_input [8i+8..8i+11] -> 1 RangeCheck for result } * num times [8n+4..8n+7] -> 1 RangeCheck for bound { [8n+i+8] -> -> 1 ForeignFieldAdd row } * num times [9n+8] -> 1 ForeignFieldAdd row (this is where the final result goes) [9n+9] -> 1 Zero row for bound result ]

pub fn create_single_foreign_field_add(start_row: usize) -> (usize, Vec<Self>)

Create a single foreign field addition gate without integrated range checks Inputs starting row Outputs tuple (next_row, circuit_gates) where next_row - next row after this gate circuit_gates - vector of circuit gates comprising this gate

pub fn extend_single_foreign_field_add(
gates: &mut Vec<Self>,
curr_row: &mut usize
)

Create foreign field addition gate by extending the existing gates

pub fn verify_foreign_field_add<G: KimchiCurve<ScalarField = F>>(
&self,
_: usize,
witness: &[Vec<F>; 15],
index: &ProverIndex<G>
) -> CircuitGateResult<()>

Verifies the foreign field addition gadget

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_foreign_field_mul(start_row: usize) -> (usize, Vec<Self>)

Create foreign field multiplication gate Inputs the starting row Outputs tuple (next_row, circuit_gates) where next_row - next row after this gate circuit_gates - vector of circuit gates comprising this gate

pub fn extend_foreign_field_mul(gates: &mut Vec<Self>, curr_row: &mut usize)

Create foreign field multiplication gate by extending the existing gates

pub fn verify_foreign_field_mul<G: KimchiCurve<ScalarField = F>>(
&self,
_: usize,
witness: &[Vec<F>; 15],
index: &ProverIndex<G>
) -> CircuitGateResult<()>

impl<F: PrimeField> CircuitGate<F>

pub fn verify_generic(
&self,
row: usize,
witness: &[Vec<F>; 15],
public: &[F]
) -> Result<(), String>

verifies that the generic gate constraints are solved by the witness

Errors

Will give error if self.typ is not GateType::Generic.

impl<F: PrimeField> CircuitGate<F>

pub fn create_generic(wires: GateWires, c: [F; 10]) -> Self

This allows you to create two generic gates that will fit in one row, check Self::create_generic_gadget for a better to way to create these gates.

pub fn create_generic_gadget(
wires: GateWires,
gate1: GenericGateSpec<F>,
gate2: Option<GenericGateSpec<F>>
) -> Self

This allows you to create two generic gates by passing the desired gate1 and gate2 as two GenericGateSpec.

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_poseidon(wires: GateWires, coeffs: [[F; 3]; 5]) -> Self

pub fn create_poseidon_gadget(
row: usize,
first_and_last_row: [GateWires; 2],
round_constants: &[Vec<F>]
) -> (Vec<Self>, usize)

create_poseidon_gadget(row, first_and_last_row, round_constants) creates an entire set of constraint for a Poseidon hash. For that, you need to pass:

the index of the first row

the first and last rows’ wires (because they are used in the permutation)

the round constants The function returns a set of gates, as well as the next pointer to the circuit (next empty absolute row)

pub fn verify_poseidon<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

Checks if a witness verifies a poseidon gate

Errors

Will give error if self.typ is not Poseidon gate, or state does not match after permutation.

pub fn ps(&self) -> F

pub fn rc(&self) -> [[F; 3]; 5]

round constant that are relevant for this specific gate

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_multi_range_check(start_row: usize) -> (usize, Vec<Self>)

Create range check gate for constraining three 88-bit values. Inputs the starting row Outputs tuple (next_row, circuit_gates) where next_row - next row after this gate circuit_gates - vector of circuit gates comprising this gate

pub fn extend_multi_range_check(gates: &mut Vec<Self>, curr_row: &mut usize)

Create foreign field muti-range-check gadget by extending the existing gates

pub fn create_range_check(start_row: usize) -> (usize, Vec<Self>)

Create single range check gate Inputs the starting row Outputs tuple (next_row, circuit_gates) where next_row - next row after this gate circuit_gates - vector of circuit gates comprising this gate

pub fn extend_range_check(gates: &mut Vec<Self>, curr_row: &mut usize)

Create foreign field range-check gate by extending the existing gates

pub fn verify_range_check<G: KimchiCurve<ScalarField = F>>(
&self,
_: usize,
witness: &[Vec<F>; 15],
index: &ProverIndex<G>
) -> CircuitGateResult<()>

Verify the witness against a range check (related) circuit gate

The following verification checks are performed

Constraint checks for circuit gates matching the self.typ kind Circuit gates used by the range check gate are: RangeChange0 and RangeCheck1

Permutation argument checks for copied cells / wiring

Plookup checks for any lookups defined

Errors

Will give error if self.typ is invalid GateType.

Panics

Will panic if padding_length is None.

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_cairo_claim(wires: GateWires) -> Self

This function creates a CairoClaim gate

pub fn create_cairo_instruction(wires: GateWires) -> Self

This function creates a CairoInstruction gate

pub fn create_cairo_flags(wires: GateWires) -> Self

This function creates a CairoFlags gate

pub fn create_cairo_transition(wires: GateWires) -> Self

This function creates a CairoTransition gate

pub fn create_cairo_gadget(row: usize, num: usize) -> (Vec<Self>, usize)

Gadget generator of the whole cairo circuits from an absolute row and number of instructions Returns a vector of gates, and the next available row after the gadget

pub fn verify_cairo_gate<G: KimchiCurve<ScalarField = F>>(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

verifies that the Cairo gate constraints are solved by the witness depending on its type

Errors

Will give error if constraint evaluation is invalid.

Panics

Will panic if constraint linearization fails.

impl<F: PrimeField> CircuitGate<F>

pub fn create_vbmul(wires: &[GateWires; 2]) -> Vec<Self>

pub fn verify_vbmul(
&self,
_row: usize,
_witness: &[Vec<F>; 15]
) -> Result<(), String>

Verify the GateType::VarBaseMul(TODO)

Errors

TODO

pub fn vbmul(&self) -> F

impl<F: PrimeField + SquareRootField> CircuitGate<F>

pub fn create_xor(new_row: usize, bits: usize) -> (usize, Vec<Self>)

Creates a XOR gadget for bits length Includes:

num_xors Xor16 gates

1 Generic gate to constrain the final row to be zero with itself Outputs tuple (next_row, circuit_gates) where

next_row : next row after this gate

gates : vector of circuit gates comprising this gate

pub fn verify_xor<G: KimchiCurve<ScalarField = F>>(
&self,
_: usize,
witness: &[Vec<F>; 15],
index: &ProverIndex<G>
) -> CircuitGateResult<()>

Verifies the xor gadget