Module kimchi::circuits::gates::endosclmul

This module implements group endomorphism optimised variable base scalar multiplication custom Plonk constraints.

EVBSM gate constraints:

b1*(b1-1) = 0

b2*(b2-1) = 0

b3*(b3-1) = 0

b4*(b4-1) = 0

((1 + (endo - 1) * b2) * xt - xp) * s1 = (2*b1-1)*yt - yp

(2xp – s1^2 + (1 + (endo - 1) * b2) * xt) * ((xp – xr) * s1 + yr + yp) = (xp – xr) * 2yp

(yr + yp)^2 = (xp – xr)^2 * (s1^2 – (1 + (endo - 1) * b2) * xt + xr)

((1 + (endo - 1) * b2) * xt - xr) * s3 = (2*b3-1)*yt - yr

(2xr – s3^2 + (1 + (endo - 1) * b4) * xt) * ((xr – xs) * s3 + ys + yr) = (xr – xs) * 2yr

(ys + yr)^2 = (xr – xs)^2 * (s3^2 – (1 + (endo - 1) * b4) * xt + xs)

n_next = 16n + 8b1 + 4b2 + 2b3 + b4

The constraints above are derived from the following EC Affine arithmetic equations:

(xq1 - xp) * s1 = yq1 - yp

(2xp – s1^2 + xq1) * ((xp – xr) * s1 + yr + yp) = (xp – xr) * 2yp

(yr + yp)^2 = (xp – xr)^2 * (s1^2 – xq1 + xr)

(xq2 - xr) * s3 = yq2 - yr

(2xr – s3^2 + xq2) * ((xr – xs) * s3 + ys + yr) = (xr – xs) * 2yr

(ys + yr)^2 = (xr – xs)^2 * (s3^2 – xq2 + xs)