kimchi::circuits::lookup::constraints

Function zk_patch

pub fn zk_patch<R: Rng + ?Sized, F: FftField>(
    e: Vec<F>,
    d: Radix2EvaluationDomain<F>,
    zk_rows: usize,
    rng: &mut R,
) -> Evaluations<F, Radix2EvaluationDomain<F>>

Pad with zeroes and then add 3 random elements in the last two rows for zero knowledge.

Panics

Will panic if evaluation and domain length do not meet the requirement.