kimchi::circuits::lookup::lookups

Type Alias JointLookupSpec

pub type JointLookupSpec<F> = JointLookup<SingleLookup<F>, LookupTableID>;

A spec for checking that the given vector belongs to a vector-valued lookup table, where the components of the vector are computed from a linear combination of locally-accessible cells.

Aliased Type

struct JointLookupSpec<F> {
    pub table_id: LookupTableID,
    pub entry: Vec<SingleLookup<F>>,
}

Fields
table_id: LookupTableID

The ID for the table associated with this lookup. Positive IDs are intended to be used for the fixed tables associated with individual gates, with negative IDs reserved for gates defined by the particular constraint system to avoid accidental collisions.

entry: Vec<SingleLookup<F>>