Type Alias kimchi::circuits::lookup::lookups::JointLookupValue

pub type JointLookupValue<F> = JointLookup<F, F>;

A concrete value or representation of a lookup.

Aliased Type

struct JointLookupValue<F> {
    pub table_id: F,
    pub entry: Vec<F>,
}

Fields
table_id: F

The ID for the table associated with this lookup. Positive IDs are intended to be used for the fixed tables associated with individual gates, with negative IDs reserved for gates defined by the particular constraint system to avoid accidental collisions.

entry: Vec<F>
Implementations

impl<F: Zero + One + Clone + Neg<Output = F> + From<u64>> JointLookupValue<F>

pub fn evaluate(&self, joint_combiner: &F, table_id_combiner: &F) -> F

Evaluate the combined value of a joint-lookup.