Function kimchi::circuits::polynomials::endomul_scalar::constraint

pub fn constraint<F: Field>(alphas: impl Iterator<Item = usize>) -> E<F>

The constraint for the endomul scalar computation

Each row corresponds to 8 iterations of the inner loop in “algorithm 2” on page 29 of this paper.

The state of the algorithm that’s updated across iterations of the loop is (a, b). It’s clear from that description of the algorithm that an iteration of the loop can be written as

(a, b, i) ->
  ( 2 * a + c_func(r_{2 * i}, r_{2 * i + 1}),
    2 * b + d_func(r_{2 * i}, r_{2 * i + 1}) )

for some functions c_func and d_func. If one works out what these functions are on every input (thinking of a two bit input as a number in {0, 1, 2, 3}), one finds they are given by

c_func(x), defined by

c_func(0) = 0

c_func(1) = 0

c_func(2) = -1

c_func(3) = 1

d_func(x), defined by

d_func(0) = -1

d_func(1) = 1

d_func(2) = 0

d_func(3) = 0

One can then interpolate to find polynomials that implement these functions on {0, 1, 2, 3}.

You can use sage, as

R = PolynomialRing(QQ, 'x')
c_func = R.lagrange_polynomial([(0, 0), (1, 0), (2, -1), (3, 1)])
d_func = R.lagrange_polynomial([(0, -1), (1, 1), (2, 0), (3, 0)])

Then, c_func is given by

2/3*x^3 - 5/2*x^2 + 11/6*x

and d_func is given by

2/3*x^3 - 7/2*x^2 + 29/6*x - 1 = c_func + (-x^2 + 3x - 1)

We lay it out as

0    1    2    3    4    5    6    7    8    9    10   11   12   13   14
n0   n8   a0   b0   a8   b8   x0   x1   x2   x3   x4   x5   x6   x7

where each xi is a two bit “crumb”.

We also use a polynomial to check that each xi is indeed in {0, 1, 2, 3},

crumb(x)
= x (x - 1) (x - 2) (x - 3)
= x^4 - 6*x^3 + 11*x^2 - 6*x
= x *(x^3 - 6*x^2 + 11*x - 6)