kimchi::circuits::polynomials::foreign_field_mul::witness

Struct ExternalChecks

pub struct ExternalChecks<F: PrimeField> {
    pub multi_ranges: Vec<[F; 3]>,
    pub limb_ranges: Vec<F>,
    pub compact_multi_ranges: Vec<[F; 2]>,
    pub bounds: Vec<[F; 3]>,
    pub high_bounds: Vec<F>,
}

Track external check witness data

Fields
multi_ranges: Vec<[F; 3]>limb_ranges: Vec<F>compact_multi_ranges: Vec<[F; 2]>bounds: Vec<[F; 3]>high_bounds: Vec<F>
Implementations

impl<F: PrimeField> ExternalChecks<F>

pub fn add_bound_check(&mut self, limbs: &[F; 3])

Track a bound check

pub fn add_high_bound_computation(&mut self, limb: &F)

Track a high bound computation

pub fn add_limb_check(&mut self, limb: &F)

Track a limb-range-check

pub fn add_multi_range_check(&mut self, limbs: &[F; 3])

Track a multi-range-check

pub fn add_compact_multi_range_check(&mut self, limbs: &[F; 2])

Track a compact-multi-range-check

pub fn extend_witness_multi_range_checks(&mut self, witness: &mut [Vec<F>; 15])

Extend the witness with external multi range_checks

pub fn extend_witness_compact_multi_range_checks( &mut self, witness: &mut [Vec<F>; 15], )

Extend the witness with external compact multi range_checks

pub fn extend_witness_limb_checks(&mut self, witness: &mut [Vec<F>; 15])

Extend the witness with external compact multi range_checks

pub fn extend_witness_bound_addition( &mut self, witness: &mut [Vec<F>; 15], foreign_field_modulus: &[F; 3], )

Extend the witness with external bound addition as foreign field addition

pub fn extend_witness_high_bounds_computation( &mut self, witness: &mut [Vec<F>; 15], foreign_field_modulus: &BigUint, )

Extend the witness with external high bounds additions as double generic gates