kimchi::circuits::polynomials

Module generic

This module implements the double generic gate.

Modules

testing

Structs

Generic

Implementation of the Generic gate

Enums

GenericGateSpec

The different type of computation that are possible with a generic gate. This type is useful to create a generic gate via the CircuitGate::create_generic_gadget function.

Constants

CONSTRAINTS

Number of constraints produced by the gate.

DOUBLE_GENERIC_COEFFS

The double generic gate actually contains two generic gates.

DOUBLE_GENERIC_REGISTERS

Number of generic of registers by a double generic gate.

GENERIC_COEFFS

Number of coefficients used by a single generic gate Three are used for the registers, one for the multiplication, and one for the constant.

GENERIC_REGISTERS

Number of generic of registers by a single generic gate