Module kimchi::circuits::polynomials::lookup

This source file implements the arithmetization of plookup constraints

Because of our ZK-rows, we can’t do the trick in the plookup paper of wrapping around to enforce consistency between the sorted lookup columns.

Instead, we arrange the LookupSorted table into columns in a snake-shape.

Like so, _ _ | | | | | | | | | | || || |

or, imagining the full sorted array is [ s0, …, s8 ], like

s0 s4 s4 s8 s1 s3 s5 s7 s2 s2 s6 s6

So the direction (“increasing” or “decreasing” (relative to LookupTable) is if i % 2 = 0 { Increasing } else { Decreasing }

Then, for each i < max_lookups_per_row, if i % 2 = 0, we enforce that the last element of LookupSorted(i) = last element of LookupSorted(i + 1), and if i % 2 = 1, we enforce that the first element of LookupSorted(i) = first element of LookupSorted(i + 1)

Overview of the protocol

We have our initial table lookup_table, with our desired values listed.

We have the implicit table lookups(witness) representing the values looked up in each row of the witness.

This table is initially variable-width, where some rows have no lookups, and others have several.

We explicitly compute this table, and where the width for a particular row is less than the maximum width, we insert a ‘dummy’ lookup value as many times as we need to to give every row the same number of lookups.

We’ll call this padded table witness_lookups.

We want to generate a sorted_table that contains every entry from the concatenated table lookup_table||witness_lookups, where values are in the same order as lookup_table, with all duplicates placed next to each other.

There’s an edge case around duplicate values in the lookup_table itself: these should appear in sorted_table at least once each time they appeared in the lookup_table.

This ensures that, for any beta and for each i, the pair lookup_table[i] + beta * lookup_table[i+1] corresponds to some distinct j such that sorted_table[j] + beta * sorted_table[j+1].

For all other values of j, sorted_table[j] = sorted_table[j+1]: since we’ve dealt with all of the ‘different’ pairs corresponding from moving from one value in lookup_table to the next, the only remaining pairs are those corresponding to the duplicates provided by the lookups in witness_lookups.

For example, if lookup_table is [0, 1, 2, 3, 4, 5] and witness_lookups is [0, 0, 0, 2, 2, 4], then sorted_table is [0, 0, 0, 0, 1, 2, 2, 2, 3, 4, 4, 5], and the differences are [(0, 0), (0, 0), (0, 0), (0, 1), (1, 2), (2, 2), (2, 2), (2, 3), (3, 4), (4, 4), (4, 5)]. The entries where the pairs are different are those that match with the lookup_table, and the equal pairs can be paired with the witness_lookups. This sorted_table is computed by the sorted function.

in order to check the multiset inclusion, we calculate the product over our sorted table: gamma * (1 + beta) + sorted_table[i] + beta * sorted_table[i+1]

again, when the adjacent terms sorted_table[i] and sorted_table[i+1] are equal, this simplifies to (gamma + sorted_table[i]) * (1 + beta)

when they are different, there is some j such that it equals gamma * (1 + beta) + lookup_table[i] + beta * lookup_table[i+1]

using the example above, this becomes

    gamma * (1 + beta) + 0 + beta * 0
  * gamma * (1 + beta) + 0 + beta * 0
  * gamma * (1 + beta) + 0 + beta * 0
  * gamma * (1 + beta) + 0 + beta * 1
  * gamma * (1 + beta) + 1 + beta * 2
  * gamma * (1 + beta) + 2 + beta * 2
  * gamma * (1 + beta) + 2 + beta * 2
  * gamma * (1 + beta) + 2 + beta * 3
  * gamma * (1 + beta) + 3 + beta * 4
  * gamma * (1 + beta) + 4 + beta * 4
  * gamma * (1 + beta) + 4 + beta * 5

which we can simplify to

    (gamma + 0) * (1 + beta)
  * (gamma + 0) * (1 + beta)
  * (gamma + 0) * (1 + beta)
  * gamma * (1 + beta) + 0 + beta * 1
  * gamma * (1 + beta) + 1 + beta * 2
  * (gamma + 2) * (1 + beta)
  * (gamma + 2) * (1 + beta)
  * gamma * (1 + beta) + 2 + beta * 3
  * gamma * (1 + beta) + 3 + beta * 4
  * (gamma + 4) * (1 + beta)
  * gamma * (1 + beta) + 4 + beta * 5

because we said before that each pair corresponds to either a pair in the lookup_table or a duplicate from the witness_table, the product over the sorted table should equal the product of gamma * (1 + beta) + lookup_table[i] + beta * lookup_table[i+1] multiplied by the product of (gamma + witness_table[i]) * (1 + beta), since each term individually cancels out.

using the example above, the lookup_table terms become

    gamma * (1 + beta) + 0 + beta * 1
  * gamma * (1 + beta) + 1 + beta * 2
  * gamma * (1 + beta) + 2 + beta * 3
  * gamma * (1 + beta) + 3 + beta * 4
  * gamma * (1 + beta) + 4 + beta * 5

and the witness_table terms become

    (gamma + 0) * (1 + beta)
  * (gamma + 0) * (1 + beta)
  * (gamma + 0) * (1 + beta)
  * (gamma + 2) * (1 + beta)
  * (gamma + 2) * (1 + beta)
  * (gamma + 4) * (1 + beta)

There is some nuance around table lengths; for example, notice that witness_table need not be the same length as lookup_table (and indeed is not in our implementation, due to multiple lookups per row), and that sorted_table will always be longer than lookup_table, which is where we require ‘snakifying’ to check consistency. Happily, we don’t have to perform snakifying on witness_table, because its contribution above only uses a single term rather than a pair of terms.

Structs

CombinedEntry

UncombinedEntry

Constants

CONSTRAINTS

Number of constraints produced by the argument.

ZK_ROWS

The number of random values to append to columns for zero-knowledge.

Traits

Entry

Functions

aggregation

Computes the aggregation polynomial for maximum n lookups per row, whose kth entry is the product of terms

constraints

Specifies the lookup constraints as expressions.

sorted

Computes the sorted lookup tables required by the lookup argument.

verify

Checks that all the lookup constraints are satisfied.

zk_patch

Pad with zeroes and then add 3 random elements in the last two rows for zero knowledge.