kimchi::circuits::polynomials::rot

Function extend_rot

pub fn extend_rot<F: PrimeField>(
    witness: &mut [Vec<F>; 15],
    word: u64,
    rot: u32,
    side: RotMode,
)

Extends the rot rows to the full witness Input

witness: full witness of the circuit

word: 64-bit word to be rotated

rot: rotation offset

side: side of the rotation, either left or right

Warning:

don’t forget to include a public input row with zero value