Module kimchi::circuits::polynomials::xor

This module includes the definition of the XOR gadget for 64, 32, and 16 bits, the definition of the constraints of the Xor16 circuit gate, and the code for witness generation for the XOR gadget.

Structs

Xor16

Constants

GATE_COUNT

Functions

circuit_gate_constraint_count

Number of constraints for a given xor gate type

circuit_gate_constraints

Get combined constraints for a given xor circuit gate type

circuit_gates

Get vector of xor circuit gate types

combined_constraints

Get the combined constraints for all xor circuit gate types

create

Create a keccak Xor for up to 128 bits Input: first input and second input

extend_xor_rows

Extends the xor rows to the full witness

lookup_table

Get the xor lookup table

num_xors

Returns the number of XOR rows needed for inputs of usize bits