Trait kimchi::folding::FoldingConfig

pub trait FoldingConfig: Clone + Debug + Eq + Hash + 'static {
    type Column: FoldingColumnTrait + Debug + Eq + Hash;
    type Challenge: Clone + Copy + Debug + Eq + Hash;
    type Curve: CommitmentCurve;
    type Srs: SRS<Self::Curve>;
    type Sponge: Sponge<Self::Curve>;
    type Instance: Instance<Self::Curve>;
    type Witness: Witness<Self::Curve>;
    type Structure;
    type Env: FoldingEnv<<Self::Curve as AffineCurve>::ScalarField, Self::Instance, Self::Witness, Self::Column, Self::Challenge, Structure = Self::Structure>;

    // Required method
    fn rows() -> usize;
}

Required Associated Types

type Column: FoldingColumnTrait + Debug + Eq + Hash

type Challenge: Clone + Copy + Debug + Eq + Hash

The type of an abstract challenge that can be found in the expressions provided as constraints.

type Curve: CommitmentCurve

The target curve used by the polynomial commitment

type Srs: SRS<Self::Curve>

type Sponge: Sponge<Self::Curve>

FIXME: use Sponge from kimchi The sponge used to create challenges

type Instance: Instance<Self::Curve>

For Plonk, it will be the commitments to the polynomials and the challenges

type Witness: Witness<Self::Curve>

For PlonK, it will be the polynomials in evaluation form that we commit to, i.e. the columns. In the generic prover/verifier, it would be kimchi_msm::witness::Witness.

type Structure

type Env: FoldingEnv<<Self::Curve as AffineCurve>::ScalarField, Self::Instance, Self::Witness, Self::Column, Self::Challenge, Structure = Self::Structure>

Required Methods

fn rows() -> usize

Return the size of the circuit, i.e. the number of rows

Object Safety

This trait is not object safe.