Trait kimchi::folding::FoldingEnv

pub trait FoldingEnv<F, I, W, Col, Chal, S> {
    type Structure;

    // Required methods
    fn new(
        structure: &Self::Structure,
        instances: [&I; 2],
        witnesses: [&W; 2]
    ) -> Self;
    fn zero_vec(&self) -> Vec<F>;
    fn col(&self, col: Col, curr_or_next: CurrOrNext, side: Side) -> &Vec<F>;
    fn lagrange_basis(&self, i: usize) -> &Vec<F>;
    fn challenge(&self, challenge: Chal, side: Side) -> F;
    fn alpha(&self, i: usize, side: Side) -> F;
    fn selector(&self, s: &S, side: Side) -> &Vec<F>;
}

Describe a folding environment. The type parameters are:

F: The field of the circuit/computation

I: The instance type, i.e the public inputs

W: The type of the witness, i.e. the private inputs

Col: The type of the column

Chal: The type of the challenge

Required Associated Types

type Structure

Structure which could be storing useful information like selectors, etc.

Required Methods

fn new( structure: &Self::Structure, instances: [&I; 2], witnesses: [&W; 2] ) -> Self

Creates a new environment storing the structure, instances and witnesses.

fn zero_vec(&self) -> Vec<F>

Returns a vector of zeros with the same length as the number of rows in the circuit.

fn col(&self, col: Col, curr_or_next: CurrOrNext, side: Side) -> &Vec<F>

Returns the evaluations of a given column witness at omega or zeta*omega.

fn lagrange_basis(&self, i: usize) -> &Vec<F>

Returns the evaluations of the i-th Lagrangian term.

fn challenge(&self, challenge: Chal, side: Side) -> F

Obtains a given challenge from the expanded instance for one side. The challenges are stored inside the instances structs.

fn alpha(&self, i: usize, side: Side) -> F

Computes the i-th power of alpha for a given side. Folding itself will provide us with the alpha value.

fn selector(&self, s: &S, side: Side) -> &Vec<F>

similar to col(), but folding may ask for a dynamic selector directly instead of just column that happens to be a selector

Object Safety

This trait is not object safe.