Trait kimchi::plonk_sponge::FrSponge

pub trait FrSponge<Fr: Field> {
    // Required methods
    fn new(p: &'static ArithmeticSpongeParams<Fr>) -> Self;
    fn absorb(&mut self, x: &Fr);
    fn absorb_multiple(&mut self, x: &[Fr]);
    fn challenge(&mut self) -> ScalarChallenge<Fr>;
    fn digest(self) -> Fr;
    fn absorb_evaluations(
        &mut self,
        e: &ProofEvaluations<PointEvaluations<Vec<Fr>>>
    );
}

Abstracts a sponge that operates on the scalar field of an elliptic curve. Unlike the FqSponge it cannot absorb or digest base field elements. However, the FqSponge can also operate on the scalar field by the means of a specific encoding technique.

Required Methods

fn new(p: &'static ArithmeticSpongeParams<Fr>) -> Self

Creates a new Fr-Sponge.

fn absorb(&mut self, x: &Fr)

Absorbs the field element into the sponge.

fn absorb_multiple(&mut self, x: &[Fr])

Absorbs a slice of field elements into the sponge.

fn challenge(&mut self) -> ScalarChallenge<Fr>

Creates a [ScalarChallenge] by squeezing the sponge.

fn digest(self) -> Fr

Consumes the sponge and returns the current digest, by squeezing.

fn absorb_evaluations( &mut self, e: &ProofEvaluations<PointEvaluations<Vec<Fr>>> )

Absorbs the given evaluations into the sponge.

Object Safety

This trait is not object safe.