Module kimchi::precomputed_srs

To prover and verify proofs you need a Structured Reference String (SRS). The generation of this SRS is quite expensive, so we provide a pre-generated SRS in this repo. Specifically, two of them, one for each pasta curve.

We generate the SRS within the test in this module. If you modify the SRS, you will need to regenerate the SRS by passing the SRS_OVERWRITE env var.

Structs

TestSRS

A clone of the SRS struct that is used for serialization, in a test-optimised way.

Enums

StoredSRSType

We store several different types of SRS objects. This enum parameterizes them.

Constants

SERIALIZED_SRS_SIZE

The size of the SRS that we serialize.

Functions

get_srs

Obtains an SRS for a specific curve from disk. Panics if the SRS does not exists.

get_srs_generic

Generic SRS getter fuction.

get_srs_test

Obtains a Test SRS for a specific curve from disk. Panics if the SRS does not exists.