kimchi

Module proof

This module implements the data structures of a proof.

Modules

caml

Structs

LookupCommitments

Commitments linked to the lookup feature

PointEvaluations

Evaluations of a polynomial at 2 points

ProofEvaluations

Polynomial evaluations contained in a ProverProof.

ProverCommitments

All the commitments that the prover creates as part of the proof.

ProverProof

The proof that the prover creates from a ProverIndex and a witness.

RecursionChallenge

A struct to store the challenges inside a ProverProof