Struct kimchi::proof::ProofEvaluations

pub struct ProofEvaluations<Evals> {
    pub w: [Evals; 15],
    pub z: Evals,
    pub s: [Evals; 6],
    pub coefficients: [Evals; 15],
    pub lookup: Option<LookupEvaluations<Evals>>,
    pub generic_selector: Evals,
    pub poseidon_selector: Evals,
}

Polynomial evaluations contained in a ProverProof.

Chunked evaluations Field is instantiated with vectors with a length that equals the length of the chunk

Non chunked evaluations Field is instantiated with a field, so they are single-sized#serde_as

Fields
w: [Evals; 15]

witness polynomials

z: Evals

permutation polynomial

s: [Evals; 6]

permutation polynomials (PERMUTS-1 evaluations because the last permutation is only used in commitment form)

coefficients: [Evals; 15]

coefficient polynomials

lookup: Option<LookupEvaluations<Evals>>

lookup-related evaluations

generic_selector: Evals

evaluation of the generic selector polynomial

poseidon_selector: Evals

evaluation of the poseidon selector polynomial

Implementations

impl<Eval> ProofEvaluations<Eval>

pub fn map<Eval2, FN: Fn(Eval) -> Eval2>(self, f: &FN) -> ProofEvaluations<Eval2>

pub fn map_ref<Eval2, FN: Fn(&Eval) -> Eval2>(
&self,
f: &FN
) -> ProofEvaluations<Eval2>

impl<F> ProofEvaluations<F>

pub fn transpose<const N: usize>(
evals: [&ProofEvaluations<F>; N]
) -> ProofEvaluations<[&F; N]>

Transpose the ProofEvaluations.

Panics

Will panic if ProofEvaluation is None.

impl<F: Zero + Copy> ProofEvaluations<PointEvaluations<F>>

pub fn dummy_with_witness_evaluations(
curr: [F; 15],
next: [F; 15]
) -> ProofEvaluations<PointEvaluations<F>>

impl<F: FftField> ProofEvaluations<PointEvaluations<Vec<F>>>

pub fn combine(
&self,
pt: &PointEvaluations<F>
) -> ProofEvaluations<PointEvaluations<F>>

impl<F> ProofEvaluations<F>

pub fn get_column(&self, col: Column) -> Option<&F>