kimchi

Module prover

This module implements prover’s zk-proof primitive.

Modules

caml

internal_traces

Macros

check_constraint 🔒

Helper to quickly test if a witness satisfies a constraint

Structs

LookupContext 🔒

Contains variables needed for lookup in the prover algorithm.

Type Aliases

Result 🔒

The result of a proof creation or verification.