Module kimchi::snarky

Snarky is the front end to kimchi, allowing users to write their own programs and convert them to kimchi circuits.

See the Mina book for more detail about this implementation.

See the tests.rs file for examples of how to use snarky.

Modules

api

The main interface to using Snarky.

asm

An ASM-like language to print a human-friendly version of a circuit.

boolean

The Boolean type is a snarky type that represents a boolean variable.

constants

Constants used for poseidon.

constraint_system

The backend used by Snarky, gluing snarky to kimchi. This module holds the actual logic that constructs the circuit using kimchi’s gates, as well as the logic that constructs the permutation, and the symbolic execution trace table (both for compilation and at runtime).

cvar

errors

poseidon

Functions associated to the Poseidon hash function.

prelude

A handy module that you can import the content of to easily use snarky.

runner

The circuit-generation and witness-generation logic.

snarky_type

The SnarkyType trait is a useful trait that allows us to define our own snarky variables on top of FieldVar. Without it, we’d be limited to using the FieldVar type directly, handling and returning arrays of FieldVars all the time. So this is useful for the same reason that programming languages allow users to define their own types, instead of relying only on usize.

union_find