Module kimchi::snarky::constraint_system

The backend used by Snarky, gluing snarky to kimchi. This module holds the actual logic that constructs the circuit using kimchi’s gates, as well as the logic that constructs the permutation, and the symbolic execution trace table (both for compilation and at runtime).

Structs

BasicInput

EcAddCompleteInput

EcEndoscaleInput

EndoscaleRound

EndoscaleScalarRound

PoseidonInput

ScaleRound

SnarkyConstraintSystem

The constraint system.

Enums

BasicSnarkyConstraint

KimchiConstraint

A PLONK constraint (or gate) can be Basic, Poseidon, EcAddComplete, EcScale, EcEndoscale, or EcEndoscalar.

Traits

SnarkyCvar

Functions

canonicalize