kimchi::circuits::argument

Trait Argument

pub trait Argument<F: PrimeField> {
    const ARGUMENT_TYPE: ArgumentType;
    const CONSTRAINTS: u32;

    // Required method
    fn constraint_checks<T: ExprOps<F, BerkeleyChallengeTerm>>(
        env: &ArgumentEnv<F, T>,
        cache: &mut Cache,
    ) -> Vec<T>;

    // Provided methods
    fn constraints(cache: &mut Cache) -> Vec<E<F>>  { ... }
    fn combined_constraints(alphas: &Alphas<F>, cache: &mut Cache) -> E<F> { ... }
}

The interface for a minimal argument implementation.

Required Associated Constants

const ARGUMENT_TYPE: ArgumentType

The type of constraints that this will produce. This is important to enforce that we don’t combine the constraints with powers of alpha that collide with other mutually inclusive arguments.

const CONSTRAINTS: u32

The number of constraints created by the argument.

Required Methods

fn constraint_checks<T: ExprOps<F, BerkeleyChallengeTerm>>( env: &ArgumentEnv<F, T>, cache: &mut Cache, ) -> Vec<T>

Constraints for this argument

Provided Methods

fn constraints(cache: &mut Cache) -> Vec<E<F>>

Returns the set of constraints required to prove this argument.

fn combined_constraints(alphas: &Alphas<F>, cache: &mut Cache) -> E<F>

Returns constraints safely combined via the passed combinator.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.