Trait kimchi::circuits::argument::Argument

pub trait Argument<F: FftField> {
    const ARGUMENT_TYPE: ArgumentType;
    const CONSTRAINTS: u32;

    fn constraints() -> Vec<E<F>>;

    fn combined_constraints(alphas: &Alphas<F>) -> E<F> { ... }
}

The interface for a minimal argument implementation.

Associated Constants

const ARGUMENT_TYPE: ArgumentType

The type of constraints that this will produce. This is important to enforce that we don’t combine the constraints with powers of alpha that collide with other mutually inclusive arguments.

const CONSTRAINTS: u32

The number of constraints created by the argument.

Required methods

fn constraints() -> Vec<E<F>>

Returns the set of constraints required to prove this argument.

Provided methods

fn combined_constraints(alphas: &Alphas<F>) -> E<F>

Returns constraints safely combined via the passed combinator.