Trait kimchi::circuits::expr::constraints::ExprOps

pub trait ExprOps<F>: Add<Output = Self> + Sub<Output = Self> + Neg<Output = Self> + Mul<Output = Self> + AddAssign<Self> + MulAssign<Self> + Clone + Zero + One + From<u64> + Displaywhere
    Self: Sized,{
    fn double(&self) -> Self;
    fn square(&self) -> Self;
    fn pow(&self, p: u64) -> Self;
    fn boolean(&self) -> Self;
    fn literal(x: F) -> Self;
    fn witness(row: CurrOrNext, col: usize, env: Option<&ArgumentData<F>>) -> Self;
    fn coeff(col: usize, env: Option<&ArgumentData<F>>) -> Self;
    fn constant(expr: ConstantExpr<F>, env: Option<&ArgumentData<F>>) -> Self;
    fn cache(&self, cache: &mut Cache) -> Self;
}

This trait defines a common arithmetic operations interface that can be used by constraints. It allows us to reuse constraint code for witness computation.

Required Methods

fn double(&self) -> Self

Double the value

fn square(&self) -> Self

Compute the square of this value

fn pow(&self, p: u64) -> Self

Raise the value to the given power

fn boolean(&self) -> Self

Constrain to boolean

fn literal(x: F) -> Self

Create a literal

fn witness(row: CurrOrNext, col: usize, env: Option<&ArgumentData<F>>) -> Self

fn coeff(col: usize, env: Option<&ArgumentData<F>>) -> Self

Coefficient

fn constant(expr: ConstantExpr<F>, env: Option<&ArgumentData<F>>) -> Self

Create a constant

fn cache(&self, cache: &mut Cache) -> Self

Cache item