Module kimchi::circuits::gate

This module implements Plonk constraint gate primitive.

Structs

Circuit

A circuit is specified as a series of CircuitGate.

CircuitGate

A single gate in a circuit.

Enums

CurrOrNext

A row accessible from a given row, corresponds to the fact that we open all polynomials at zeta and omega * zeta.

GateType

The different types of gates the system supports. Note that all the gates are mutually exclusive: they cannot be used at the same time on single row. If we were ever to support this feature, we would have to make sure not to re-use powers of alpha across constraints.