kimchi::circuits::gate

Enum GateType

#[repr(C)]
pub enum GateType {

    Zero = 0,
    Generic = 1,
    Poseidon = 2,
    CompleteAdd = 3,
    VarBaseMul = 4,
    EndoMul = 5,
    EndoMulScalar = 6,
    Lookup = 7,
    CairoClaim = 8,
    CairoInstruction = 9,
    CairoFlags = 10,
    CairoTransition = 11,
    RangeCheck0 = 12,
    RangeCheck1 = 13,
    ForeignFieldAdd = 14,
    ForeignFieldMul = 15,
    Xor16 = 16,
    Rot64 = 17,

}

The different types of gates the system supports. Note that all the gates are mutually exclusive: they cannot be used at the same time on single row. If we were ever to support this feature, we would have to make sure not to re-use powers of alpha across constraints.

Variants

Zero = 0

Zero gate

Generic = 1

Generic arithmetic gate

Poseidon = 2

Poseidon permutation gate

CompleteAdd = 3

Complete EC addition in Affine form

VarBaseMul = 4

EC variable base scalar multiplication

EndoMul = 5

EC variable base scalar multiplication with group endomorphim optimization

EndoMulScalar = 6

Gate for computing the scalar corresponding to an endoscaling

Lookup = 7

CairoClaim = 8

Cairo

CairoInstruction = 9

CairoFlags = 10

CairoTransition = 11

RangeCheck0 = 12

Range check

RangeCheck1 = 13

ForeignFieldAdd = 14

ForeignFieldMul = 15

Xor16 = 16

Rot64 = 17

Implementations

impl GateType

pub fn lookup_kinds() -> Vec<LookupPattern>

Which lookup-patterns should be applied on which rows.