Struct kimchi::circuits::gate::CircuitGate

pub struct CircuitGate<F: FftField> {
    pub typ: GateType,
    pub wires: GateWires,
    pub coeffs: Vec<F>,
}

A single gate in a circuit.

Fields
typ: GateType

type of the gate

wires: GateWires

gate wiring (for each cell, what cell it is wired to)

coeffs: Vec<F>

public selector polynomials that can used as handy coefficients in gates

Implementations

impl<F: FftField> CircuitGate<F>

pub fn zero(wires: GateWires) -> Self

this function creates “empty” circuit gate

pub fn verify(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

This function verifies the consistency of the wire assignements (witness) against the constraints

impl<F: FftField> CircuitGate<F>

pub fn create_endomul(wires: GateWires) -> Self

pub fn verify_endomul(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

pub fn endomul(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn create_poseidon(wires: GateWires, coeffs: [[F; 3]; 5]) -> Self

pub fn create_poseidon_gadget(
row: usize,
first_and_last_row: [GateWires; 2],
round_constants: &[Vec<F>]
) -> (Vec<Self>, usize)

create_poseidon_gadget(row, first_and_last_row, round_constants) creates an entire set of constraint for a Poseidon hash. For that, you need to pass:

the index of the first row

the first and last rows’ wires (because they are used in the permutation)

the round constants The function returns a set of gates, as well as the next pointer to the circuit (next empty absolute row)

pub fn verify_poseidon(
&self,
row: usize,
witness: &[Vec<F>; 15],
cs: &ConstraintSystem<F>
) -> Result<(), String>

Checks if a witness verifies a poseidon gate

pub fn ps(&self) -> F

pub fn rc(&self) -> [[F; 3]; 5]

round constant that are relevant for this specific gate

impl<F: FftField> CircuitGate<F>

pub fn create_vbmul(wires: &[GateWires; 2]) -> Vec<Self>

pub fn verify_vbmul(
&self,
_row: usize,
_witness: &[Vec<F>; 15]
) -> Result<(), String>

pub fn vbmul(&self) -> F

impl<F: FftField> CircuitGate<F>

pub fn verify_complete_add(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

Check the correctness of witness values for a complete-add gate.

impl<F: FftField> CircuitGate<F>

pub fn verify_endomul_scalar(
&self,
row: usize,
witness: &[Vec<F>; 15],
_cs: &ConstraintSystem<F>
) -> Result<(), String>

impl<F: FftField> CircuitGate<F>

pub fn verify_generic(
&self,
row: usize,
witness: &[Vec<F>; 15]
) -> Result<(), String>

verifies that the generic gate constraints are solved by the witness

impl<F: FftField> CircuitGate<F>

pub fn create_generic(wires: GateWires, c: [F; 10]) -> Self

This allows you to create two generic gates that will fit in one row, check [create_generic_gadget] for a better to way to create these gates.

pub fn create_generic_gadget(
wires: GateWires,
gate1: GenericGateSpec<F>,
gate2: Option<GenericGateSpec<F>>
) -> Self

This allows you to create two generic gates by passing the desired gate1 and gate2 as two GenericGateSpec.