Trait kimchi::circuits::gate::Connect

pub trait Connect {
    fn connect_cell_pair(&mut self, cell1: (usize, usize), cell2: (usize, usize));
}

Trait to connect a pair of cells in a circuit

Required Methods

fn connect_cell_pair(&mut self, cell1: (usize, usize), cell2: (usize, usize))

Connect the pair of cells specified by the cell1 and cell2 parameters cell_pre –> cell_new && cell_new –> wire_tmp

Note: This function assumes that the targeted cells are freshly instantiated with self-connections. If the two cells are transitively already part of the same permutation then this would split it.