kimchi::circuits::lookup

Module constraints

Structs

LookupConfiguration

Configuration for the lookup constraint. These values are independent of the choice of lookup values.

Constants

CONSTRAINTS

Number of constraints produced by the argument.

Functions

aggregation

Computes the aggregation polynomial for maximum n lookups per row, whose kth entry is the product of terms

constraints

Specifies the lookup constraints as expressions.

sorted

Computes the sorted lookup tables required by the lookup argument.

verify

Checks that all the lookup constraints are satisfied.

zk_patch

Pad with zeroes and then add 3 random elements in the last two rows for zero knowledge.