kimchi::circuits::lookup

Module lookups

Modules

wasm

Structs

JointLookup

A spec for checking that the given vector belongs to a vector-valued lookup table.

LocalPosition

A position in the circuit relative to a given row.

LookupFeatures

Configuration for lookup-related features in a circuit.

LookupInfo

Describes the desired lookup configuration.

LookupPatternIter

An iterator over the variants of LookupPattern

LookupPatterns

Flags for each of the built-in lookup patterns.

SingleLookup

Look up a single value in a lookup table. The value may be computed as a linear combination of locally-accessible cells.

Enums

LookupPattern

LookupTableID

The table ID associated with a particular lookup

Constants

_DERIVE_ocaml_FromValue_from_value_lifetime_FOR_LookupFeatures 🔒

_DERIVE_ocaml_FromValue_from_value_lifetime_FOR_LookupPattern 🔒

_DERIVE_ocaml_FromValue_from_value_lifetime_FOR_LookupPatterns 🔒

_DERIVE_ocaml_IntoValue_FOR_LookupFeatures 🔒

_DERIVE_ocaml_IntoValue_FOR_LookupPattern 🔒

_DERIVE_ocaml_IntoValue_FOR_LookupPatterns 🔒

Functions

max_lookups_per_row 🔒

Type Aliases

Evaluations 🔒

JointLookupSpec

A spec for checking that the given vector belongs to a vector-valued lookup table, where the components of the vector are computed from a linear combination of locally-accessible cells.

JointLookupValue

A concrete value or representation of a lookup.