kimchi::circuits::lookup

Module lookups

Modules

wasm

Structs

JointLookup

A spec for checking that the given vector belongs to a vector-valued lookup table.

LocalPosition

A position in the circuit relative to a given row.

LookupFeatures

LookupInfo

Describes the desired lookup configuration.

LookupPatternIter

An iterator over the variants of LookupPattern

LookupPatterns

Flags for each of the hard-coded lookup patterns.

SingleLookup

Look up a single value in a lookup table. The value may be computed as a linear combination of locally-accessible cells.

Enums

LookupPattern

LookupTableID

The table ID associated with a particular lookup

Type Aliases

JointLookupSpec

A spec for checking that the given vector belongs to a vector-valued lookup table, where the components of the vector are computed from a linear combination of locally-accessible cells.

JointLookupValue

A concrete value or representation of a lookup.