Module kimchi::circuits::lookup::lookups

Structs

JointLookup

A spec for checking that the given vector belongs to a vector-valued lookup table.

LocalPosition

A position in the circuit relative to a given row.

LookupInfo

Describes the desired lookup configuration.

LookupPatternIter

An iterator over the variants of Self

SingleLookup

Look up a single value in a lookup table. The value may be computed as a linear combination of locally-accessible cells.

Enums

LookupPattern

LookupTableID

The table ID associated with a particular lookup

LookupsUsed

Specifies whether a constraint system uses joint lookups. Used to make sure we squeeze the challenge joint_combiner when needed, and not when not needed.

Type Definitions

JointLookupSpec

A spec for checking that the given vector belongs to a vector-valued lookup table, where the components of the vector are computed from a linear combination of locally-accessible cells.

JointLookupValue

A concrete value or representation of a lookup.