Type Alias kimchi::circuits::lookup::lookups::JointLookupValue

pub type JointLookupValue<F> = JointLookup<F, F>;

A concrete value or representation of a lookup.

Aliased Type

struct JointLookupValue<F> {
    pub table_id: F,
    pub entry: Vec<F, Global>,
}

Fields
table_id: F

The ID for the table associated with this lookup. Positive IDs are intended to be used for the fixed tables associated with individual gates, with negative IDs reserved for gates defined by the particular constraint system to avoid accidental collisions.

entry: Vec<F, Global>
Implementations

impl<F: Zero + One + Clone + Neg<Output = F> + From<u64>> JointLookupValue<F>

pub fn evaluate(&self, joint_combiner: &F, table_id_combiner: &F) -> F

Evaluate the combined value of a joint-lookup.

impl<F: Copy> JointLookup<SingleLookup<F>, LookupTableID>

pub fn reduce<K, G: Fn(LocalPosition) -> K>( &self, eval: &G ) -> JointLookupValue<K>where K: Zero + Mul<F, Output = K> + Neg<Output = K> + From<u64>,

Reduce linear combinations in the lookup entries to a single value, resolving local positions using the given function.

pub fn evaluate<K, G: Fn(LocalPosition) -> K>( &self, joint_combiner: &K, table_id_combiner: &K, eval: &G ) -> Kwhere K: Zero + One + Clone + Mul<F, Output = K> + Neg<Output = K> + From<u64>,

Evaluate the combined value of a joint-lookup, resolving local positions using the given function.