kimchi::circuits::lookup

Module tables

Modules

caml

range_check

Range check table

xor

Structs

GateLookupTables

Enumerates the different ‘fixed’ lookup tables used by individual gates

LookupTable

A table of values that can be used for a lookup, along with the ID for the table.

Enums

GateLookupTable

Enumerates the different ‘fixed’ lookup tables used by individual gates

Constants

RANGE_CHECK_TABLE_ID

The range check table ID.

XOR_TABLE_ID

The table ID associated with the XOR lookup table.

Functions

combine_table

Same as combine_table_entry, but for an entire table. The function will panic if given an empty table (0 columns).

combine_table_entry

Let’s say we want to do a lookup in a “vector-valued” table T: Vec<[F; n]> (here I am using [F; n] to model a vector of length n).

get_table

Returns the lookup table associated to a GateLookupTable.