Module kimchi::circuits::polynomials

Modules

chacha

This module implements the cha-cha constraints, which are rather simple aside form the lookups. There are four chacha constraint types, corresponding to the four lines in each quarter round.

complete_add

This module implements a complete EC addition gate.

endomul_scalar

endosclmul

This module implements short Weierstrass curve endomorphism optimised variable base scalar multiplication custom Plonk polynomials.

generic

This module implements the double generic gate.

lookup

This source file implements the arithmetization of plookup constraints

permutation

This module implements permutation constraint polynomials.

poseidon

This module implements the Poseidon constraint polynomials.

varbasemul

This module implements short Weierstrass curve variable base scalar multiplication custom Plonk polynomials.