Module kimchi::circuits::polynomials::generic

This module implements the double generic gate, which contains two generic gates.

A generic gate is simply the 2-fan in gate specified in the vanilla PLONK protocol that allows us to do:

addition of two registers (into an output register)

or multiplication of two registers

equality of a register with a constant

In each cases, registers can also be scaled by a constant.

The layout of the gate is the following:

0	1	2	3	4	5	6	7	8	9	10	11	11	12	13	14

l1	r1	o1	l2	r2	o2

where l1, r1, and o1 (resp. l2, r2, o2) are the left, right, and output registers of the first (resp. second) generic gate.

For the selector:

0	1	2	3	4	5	6	7	8	9	10	11	11	12	13	14

l1	r1	o1	m1	c1	l2	r2	o2	m2	c2

with m1 (resp. m2) the mul selector for the first (resp. second) gate, and c1 (resp. c2) the constant selector for the first (resp. second) gate.

The polynomial looks like this:

[
  alpha1 * (w0 * coeff0 + w1 * coeff1 + w2 * coeff2 + w0 * w1 * coeff3 + coeff4) +
  alpha2 * (w3 * coeff5 + w4 * coeff6 + w5 * coeff7 + w3 w4 coeff8 + coeff9)
] * generic_selector

Modules

testing

Enums

GenericGateSpec

The different type of computation that are possible with a generic gate. This type is useful to create a generic gate via the [create_generic_gadget] function.

Constants

CONSTRAINTS

Number of constraints produced by the gate.

GENERIC_COEFFS

Number of coefficients used by a single generic gate Three are used for the registers, one for the multiplication, and one for the constant.

GENERIC_REGISTERS

Number of generic of registers by a single generic gate