kimchi::circuits::polynomials::keccak

Module constants

Constants for each witness’ index offsets and lengths

Constants

CAPACITY_IN_BYTES

The number of bytes used as a capacity in the sponge.

CHI_SHIFTS_B_LEN

CHI_SHIFTS_B_OFF

CHI_SHIFTS_SUM_LEN

CHI_SHIFTS_SUM_OFF

DIM

The dimension of the Keccak state

IOTA_STATE_G_LEN

IOTA_STATE_G_OFF

KECCAK_COLS

The number of columns the Keccak circuit uses.

PIRHO_DENSE_E_LEN

PIRHO_DENSE_E_OFF

PIRHO_DENSE_ROT_E_LEN

PIRHO_DENSE_ROT_E_OFF

PIRHO_EXPAND_ROT_E_LEN

PIRHO_EXPAND_ROT_E_OFF

PIRHO_QUOTIENT_E_LEN

PIRHO_QUOTIENT_E_OFF

PIRHO_REMAINDER_E_LEN

PIRHO_REMAINDER_E_OFF

PIRHO_SHIFTS_E_LEN

PIRHO_SHIFTS_E_OFF

QUARTERS

An element of the Keccak state is 64 bits. However, we split the state into quarters of 16 bits to represent XOR and AND using the finite field addition. See the Keccak RFC for more information.

RATE_IN_BYTES

The number of bytes that can be processed by the Keccak permutation. It is the rate of the sponge configuration.

ROUNDS

The number of rounds in the Keccak permutation

SHIFTS

SHIFTS_LEN

SPONGE_BYTES_LEN

SPONGE_BYTES_OFF

SPONGE_COLS

The number of columns the Sponge circuit uses.

SPONGE_NEW_BLOCK_LEN

SPONGE_NEW_BLOCK_OFF

SPONGE_NEW_STATE_LEN

SPONGE_NEW_STATE_OFF

SPONGE_OLD_STATE_LEN

SPONGE_OLD_STATE_OFF

SPONGE_SHIFTS_LEN

SPONGE_SHIFTS_OFF

SPONGE_XOR_STATE_LEN

SPONGE_XOR_STATE_OFF

SPONGE_ZEROS_LEN

SPONGE_ZEROS_OFF

STATE_LEN

The number of field elements used to represent the whole Keccak state.

THETA_DENSE_C_LEN

THETA_DENSE_C_OFF

THETA_DENSE_ROT_C_LEN

THETA_DENSE_ROT_C_OFF

THETA_EXPAND_ROT_C_LEN

THETA_EXPAND_ROT_C_OFF

THETA_QUOTIENT_C_LEN

THETA_QUOTIENT_C_OFF

THETA_REMAINDER_C_LEN

THETA_REMAINDER_C_OFF

THETA_SHIFTS_C_LEN

THETA_SHIFTS_C_OFF

THETA_STATE_A_LEN

THETA_STATE_A_OFF