kimchi::circuits::polynomials::not

Function extend_not_witness_checked_length

pub fn extend_not_witness_checked_length<F: PrimeField>(
    witness: &mut [Vec<F>; 15],
    input: F,
    bits: Option<usize>,
)

Extend a NOT witness for less than 255 bits (native field) Input: full witness, first input and optional bit length If bits is not provided, the negation is performed using the length of the input in bits. If bits is provided, the negation takes the maximum length between bits and that of input. Warning:

don’t forget to set a row of the witness with public input 2^bits -1 and wire it to the second input of the first Xor16 gate