Function kimchi::circuits::polynomials::poseidon::generate_witness

pub fn generate_witness<F: Field>(
    row: usize, 
    params: ArithmeticSpongeParams<F>, 
    witness_cols: &mut [Vec<F>; 15], 
    input: [F; 3]
)

generate_witness(row, params, witness_cols, input) uses a sponge initialized with params to generate a witness for starting at row row in witness_cols, and with input input.