kimchi::circuits::wires

Trait Wirable

pub trait Wirable: Sized {
    // Required methods
    fn new(row: usize) -> Self;
    fn wire(self, col: usize, to: Wire) -> Self;
}

Since we don’t have a specific type for the wires of a row, we have to implement these convenience functions through a trait.

Required Methods

fn new(row: usize) -> Self

Creates a new set of wires for a given row.

fn wire(self, col: usize, to: Wire) -> Self

Wire the cell at col to another cell (to).

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.