Module kimchi::circuits::witness

Macros

variable_map

Macro to simplify creation of layout map

variables

Macro to simplify mapping of layout variable

Structs

ConstantCell

Witness cell with constant value

CopyBitsCell

Witness cell copied from bits of another witness cell

CopyCell

Witness cell copied from another witness cell

CopyShiftCell

Witness cell copied from another cell and shifted

CrumbCell

Witness cell assigned from a variable and a crumb index See Variables for more details

VariableCell

Witness cell assigned from a variable See Variables for more details

Variables

Layout variables mapping - these values are substituted into the witness layout when creating the witness instance.

Traits

WitnessCell

Witness cell interface

Functions

init

Initialize a witness based on layout and computed variables

init_cell

Initialize a witness cell based on layout and computed variables

init_row

Initialize a witness row based on layout and computed variables