Trait kimchi::curve::KimchiCurve

pub trait KimchiCurve: CommitmentCurve {
    type OtherCurve: KimchiCurve<ScalarField = Self::BaseField, BaseField = Self::ScalarField, OtherCurve = Self>;

    // Required methods
    fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField>;
    fn endos() -> &'static (Self::BaseField, Self::ScalarField);
}

Represents additional information that a curve needs in order to be used with Kimchi

Required Associated Types

type OtherCurve: KimchiCurve<ScalarField = Self::BaseField, BaseField = Self::ScalarField, OtherCurve = Self>

The other curve that forms the cycle used for recursion

Required Methods

fn sponge_params() -> &'static ArithmeticSpongeParams<Self::ScalarField>

Provides the sponge params to be used with this curve If the params for the base field are needed, they can be obtained from KimchiCurve::OtherCurve

fn endos() -> &'static (Self::BaseField, Self::ScalarField)

Provides the coefficients for the curve endomorphism