Trait kimchi::plonk_sponge::FrSponge

pub trait FrSponge<Fr: Field> {
    fn new(p: &'static ArithmeticSpongeParams<Fr>) -> Self;
    fn absorb(&mut self, x: &Fr);
    fn absorb_multiple(&mut self, x: &[Fr]);
    fn challenge(&mut self) -> ScalarChallenge<Fr>;
    fn digest(self) -> Fr;
    fn absorb_evaluations(
        &mut self,
        e: &ProofEvaluations<PointEvaluations<Vec<Fr>>>
    );
}

Required Methods

fn new(p: &'static ArithmeticSpongeParams<Fr>) -> Self

Creates a new Fr-Sponge.

fn absorb(&mut self, x: &Fr)

Absorbs the field element into the sponge.

fn absorb_multiple(&mut self, x: &[Fr])

Absorbs a slice of field elements into the sponge.

fn challenge(&mut self) -> ScalarChallenge<Fr>

Creates a [ScalarChallenge] by squeezing the sponge.

fn digest(self) -> Fr

Consumes the sponge and returns the current digest, by squeezing.

fn absorb_evaluations(&mut self, e: &ProofEvaluations<PointEvaluations<Vec<Fr>>>)

Absorbs the given evaluations into the sponge.